// src/prisma/generated/zod/inputTypeSchemas/JamUncheckedCreateInputSchema.ts
import { z } from 'zod';
import { type Prisma } from '@prisma/client';
import { VisibilitySchema } from './VisibilitySchema';
import { TicketUncheckedCreateNestedManyWithoutJamInputSchema } from './TicketUncheckedCreateNestedManyWithoutJamInputSchema';
import { TicketConfigurationUncheckedCreateNestedManyWithoutJamInputSchema } from './TicketConfigurationUncheckedCreateNestedManyWithoutJamInputSchema';
import { JamOrganiserUncheckedCreateNestedManyWithoutJamInputSchema } from './JamOrganiserUncheckedCreateNestedManyWithoutJamInputSchema';

export const JamUncheckedCreateInputSchema: z.ZodType<Prisma.JamUncheckedCreateInput> = z.object({
  id: z.string(),
  name: z.string(),
  startDatetime: z.coerce.date(),
  endDatetime: z.coerce.date(),
  visibility: z.lazy(() => VisibilitySchema),
  summary: z.string().optional().nullable(),
  description: z.string().optional().nullable(),
  venue: z.string().optional().nullable(),
  tickets: z.lazy(() => TicketUncheckedCreateNestedManyWithoutJamInputSchema).optional(),
  ticketConfiguration: z.lazy(() => TicketConfigurationUncheckedCreateNestedManyWithoutJamInputSchema).optional(),
  organisers: z.lazy(() => JamOrganiserUncheckedCreateNestedManyWithoutJamInputSchema).optional(),
}).strict();

export default JamUncheckedCreateInputSchema;